requirements.txt
boto3
sagemaker
Pillow
streamlit
